Remove the two error outputs in the sandbox.
https://github.com/chunhualiao/freeCompilerCamp/issues/52
There were two error outputs.

1) "/bin/sh: 1: cannot create /etc/hosts: Permission denied"
As a non-root user, freecc tried to access that system file during start and got denied.
I tried to change the ownership or permission of this file in the Dockerfile. However, it's not allowed by docker, which led a failed image building.
For now, a "clear" command is inserted to ~/.profile of freecc. It clean up the error output before anyone can see it in the sandbox.
This issue will be addressed later to find a permanent solution.

2) "bash: /var/run/cwd: Permission denied"
Similarly, in the ~/.profile of freecc, it tried to write a system file and got denied. This specific command has been commented out to prevent the error.

After these two changes, Dockerfile.dind is rebuilt. Container pwd is also rebuilt based on the new image. Now the terminal is clean.